Alternative StateT representation that admits stack-safe operations for tail-recursive monads.
The benefit is that one does not need to trampoline StateT anymore
(the trampoline is built-in).

The trade-off is that iterated mapK will not be stack-safe
even for a lazy inner monad, because the mapK operation cannot be
represented lazily in this new representation. This can be justified
by the fact that mapK changes the inner monad, and there will only
be a constant number of different monads in any program, therefore
at most a constant number of mapK's should be necessary in any program.
This is similar to iterated Free.mapSuspension not being stack-safe.